+++ djangocms_alias/__init__.py
@@ -1,3 +1,3 @@
-__version__ = '1.4.1'
+__version__ = '1.5.0'
 
 default_app_config = 'djangocms_alias.apps.AliasConfig'