+++ djangocms_alias/admin.py
@@ -37,6 +37,7 @@ if djangocms_versioning_enabled:
     from .filters import UnpublishedFilter
     alias_content_admin_classes.insert(0, ExtendedVersionAdminMixin)
     alias_content_admin_list_display = ('name', 'get_category',)
+    alias_content_admin_list_filter = (SiteFilter, LanguageFilter, UnpublishedFilter)
     alias_content_admin_list_filter = (SiteFilter, LanguageFilter, 'alias__category', UnpublishedFilter)